Revert "restore the version of the tokens api"
This reverts commit 18f12d3d81997f845c250f9637f1cb1a23bb4e05.
